pyproject.toml
[project]
name = "jujube-blackpatch-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.4.0",
    "numpy>=1.20.0",
    "opencv-python>=4.5.0",
    "pillow>=8.0.0",
    "psutil>=5.8.0",
    "pyside6>=6.0.0",
    "python-socketio>=5.0.0",
    "shapely>=2.0.0",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "ultralytics>=8.0.0",
    "wmi>=1.5.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
